requirements.txt
amqp==2.4.2
baidu-aip==2.2.13.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.1
django-celery-results==1.0.4
django-cors-headers==3.0.1
django-filter==2.1.0
django-redis-sessions==0.6.1
djangorestframework==3.9.4
idna==2.8
kombu==4.5.0
Markdown==3.1.1
oauthlib==3.0.1
Pillow==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.1
qcloudsms-py==0.1.3
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.24.3
vine==1.3.0
